+++ saveState.hh
@@ -9,6 +9,6 @@
 void dumpState(const state_t &s, const std::string &filename);
 void loadState(state_t &s, const std::string &filename);
 
-void emitCodeForInitialRegisterValues(state_t &s, uint32_t pc);
+void emitCodeForInitialRegisterValues(state_t &s, uint64_t pc);
 
 #endif
